working!

+++ js/src/main/scala/quizleague/web/util/DateTimeConverters.scala
@@ -12,7 +12,7 @@ object DateTimeConverters{
   implicit def intToYear(int:Int):Year = Year of int
   implicit def stringToLocalDate(date:String):LocalDate = LocalDate parse(date)
   implicit def localDateToString(date:LocalDate):String = date.toString
-  implicit def dateToLocalDate(date:js.Date):LocalDate = LocalDate of(date.getFullYear,date.getMonth + 1,date.getDate)
+  implicit def dateToLocalDate(date:js.Date):LocalDate = LocalDate of(date.getFullYear.toInt,date.getMonth.toInt + 1,date.getDate.toInt)
   implicit def localDateToDate(date:LocalDate):js.Date = new js.Date(js.Date parse date.toString)
   implicit def dateToLocalTime(date:String):LocalTime = LocalTime parse date
   implicit def localTimeToDate(date:LocalTime):String = f"${date.getHour}%02d:${date.getMinute}%02d"
